requirements.txt
accelerate==0.23.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
brotlipy==0.7.0
cachetools==5.3.2
certifi
cffi
charset-normalizer
click==8.1.7
clip==0.2.0
cmake==3.28.3
contourpy==1.1.1
cryptography
cycler==0.12.1
decord==0.6.0
deepspeed==0.12.3
defusedxml==0.7.1
diffusers==0.22.3
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.1.3
fastapi==0.104.0
ffmpeg==1.4
ffmpy==0.3.1
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==6.1.3
gdown==4.7.1
Genshi==0.7.7
gitdb==4.0.10
GitPython==3.1.40
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
hjson==3.1.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.19.4
idna
imageio==2.27.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
invisible-watermark==0.2.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
lightning-utilities==0.10.0
lit==18.1.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mkl-fft
mkl-random
mkl-service==2.4.0
mpi4py
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.1
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
open-clip-torch==2.23.0
opencv-python==4.8.1.78
orjson==3.9.9
packaging==23.2
pandas==2.1.1
passlib==1.7.4
pathtools==0.1.2
Pillow
polib==1.2.0
protobuf==4.24.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyav==11.4.1
pycparser
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyOpenSSL
pyparsing==3.1.1
pyre-extensions==0.0.29
PySocks
python-dateutil==2.8.2
python-magic==0.4.27
python-multipart==0.0.6
python-sql==1.4.2
pytorch-lightning==2.1.2
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
relatorio==0.10.1
requests
rich==13.7.0
rpds-py==0.10.6
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.12
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==0.13.1
torchmetrics==1.2.1
torchvision==0.14.1
tornado==6.4
tqdm==4.66.1
transformers==4.27.0
triton==2.0.0
trytond==6.8.5
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3
uvicorn==0.23.2
validators==0.22.0
wandb==0.15.12
watchdog==3.0.0
wcwidth==0.2.12
websockets==11.0.3
Werkzeug==3.0.0
xformers==0.0.19
yarl==1.9.3
zipp==3.17.0